+++ src/main/java/com/ros/lmsdesktopclient/view_models/LoginViewModel.java
@@ -1,6 +1,8 @@
 package com.ros.lmsdesktopclient.view_models;
 
 import com.ros.lmsdesktopclient.models.LoginModel;
+import com.ros.lmsdesktopclient.services.ServiceFactory;
+import com.ros.lmsdesktopclient.services.service.LoginService;
 import com.ros.lmsdesktopclient.services.service_impl.LoginServiceImpl;
 import com.ros.lmsdesktopclient.view_models.commands.Command;
 import com.ros.lmsdesktopclient.view_models.commands.LoginCommand;
@@ -11,7 +13,8 @@ public class LoginViewModel {
 
     public LoginViewModel(LoginModel loginModel){
         this.loginModel = loginModel;
-        this.loginCommand = new LoginCommand(this.loginModel, new LoginServiceImpl());
+        LoginService loginService = ServiceFactory.createProxy(LoginService.class, new LoginServiceImpl());
+        this.loginCommand = new LoginCommand(this.loginModel, loginService);
     }
 
     public LoginModel getLoginModel() {
